refactor: environments in feature directory
