+++ src/lib/features/environments/environment-oss.e2e.test.ts
@@ -1,9 +1,9 @@
 import {
     type IUnleashTest,
     setupAppWithCustomConfig,
-} from '../../helpers/test-helper';
-import dbInit, { type ITestDb } from '../../helpers/database-init';
-import getLogger from '../../../fixtures/no-logger';
+} from '../../../test/e2e/helpers/test-helper';
+import dbInit, { type ITestDb } from '../../../test/e2e/helpers/database-init';
+import getLogger from '../../../test/fixtures/no-logger';
 
 let app: IUnleashTest;
 let db: ITestDb;
